requirements.txt
# Remaining requirements that should be installed if not already done so
numpy==1.16.5
pandas==0.25.1
tqdm
scipy>=1.3.1
scikit-learn>=0.21.2
# flake8
matplotlib>=3.1.1
GPUtil
setuptools
tensorflow-gpu==1.14.0
keras==2.3.1
umap
more-itertools
ipdb
# for beatgan
biosppy==0.6.1
seaborn
# for omnianomaly
tensorflow-probability==0.7.0
gast==0.2.2
git+https://github.com/korepwx/tfsnippet.git
git+https://github.com/thu-ml/zhusuan.git
# for ocsvm
thundersvm-cuda10
